requirements.txt
beautifulsoup4==4.9.3
ccxt==1.39.37
clint==0.5.1
django-decouple==2.1
flake8==3.8.4
google==3.0.0
huepy==1.2.1
ipython==7.19.0
lxml==4.6.2
matplotlib==3.3.3
mysql-connector-python==8.0.22
oauth2client==4.1.3
opencv-python==4.4.0.46
Pillow==8.0.1
psycopg2-binary==2.8.6
PyMySQL==0.10.1
PyQt5==5.15.2
PySimpleGUI==4.32.1
pytest-cov==2.10.1
pytest==6.2.1
requests==2.25.0
rows==0.4.1
scrapinghub==2.3.1
Scrapy==2.4.1
selenium==4.0.0a7
SQLAlchemy-Utils==0.36.8
Twisted==20.3.0
xmltodict==0.12.0
